refactoring backend
