+++ backend/routes/callYelp.js
@@ -2,7 +2,9 @@ const express = require('express');
 const router = express.Router();
 const yelp = require("../Yelp/YelpGet").client;
 
+//api call. 
 router.get('/callYelp', function(req, res, next) {
+    //here we use "yelp" to search our query's search term and send a JSON version of yelp's data
     yelp.search({ term: req.query.searchTerm, location: 'New York' })
         .then(function (data) {
             res.setHeader('Content-Type', 'application/json');
